+++ Makefile.am
@@ -4702,7 +4702,6 @@ EXTRA_DIST += \
 	shared/nm-test-utils-impl.c \
 	shared/nm-utils/c-list-util.c \
 	shared/nm-utils/c-list-util.h \
-	shared/nm-utils/gsystem-local-alloc.h \
 	shared/nm-utils/nm-c-list.h \
 	shared/nm-utils/nm-compat.c \
 	shared/nm-utils/nm-compat.h \
